fix timeline ownership check in addedittimeline
